requirements.txt
discord.py==2.3.2
youtube_dl==2021.12.17
spotipy==2.23.0
PyNaCl==1.5.0
python-dotenv==1.0.0
